got thread working. need to render data
